golang_tutorial/learn_by_test/hello_world/wallet.go
package main

import (
	"errors"
	"fmt"
)

type Bitcoin int

func (b Bitcoin) String() string {
	return fmt.Sprintf("%d BTC", b)
}

type Wallet struct {
	balance Bitcoin // lowercase symbol exposed only within the package
}

// these passed parameters are by-value by default
func (w *Wallet) Deposit(amount Bitcoin) {
	w.balance += amount // pointers are automatically dereferenced
}

// a global error variable as a singleton error, really?
var ErrInsufficientFunds = errors.New("cannot withdraw, insufficient funds")

func (w *Wallet) Withdraw(amount Bitcoin) error {

	if amount > w.balance {
		return ErrInsufficientFunds
	}

	w.balance -= amount
	return nil
}

func (w *Wallet) Balance() Bitcoin {
	return w.balance
}
